+++ packages/electron-updater/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## 4.3.0
 
+## 7.0.0-alpha.3
+
+### Patch Changes
+
+<details><summary>Updated 1 dependency</summary>
+
+<small>
+
+[`deafee9`](https://github.com/electron-userland/electron-builder/commit/deafee9fe5f7bccbf52e73cd0c6085e767f921ce)
+
+</small>
+
+- `builder-util-runtime@10.0.0-alpha.3`
+
+</details>
+
 ## 7.0.0-alpha.2
 
 ### Patch Changes